Move view/lang fixtures and drop WithFixtures for Fixtures
