+++ tests/Integration/Translation/Fixtures/lang/en/app.php
