+++ tests/Integration/Translation/Fixtures/lang/fr.json
