+++ tests/Integration/Translation/Fixtures/lang/fr/app.php
