+++ tests/Integration/View/BladeAnonymousComponentTest.php
@@ -11,8 +11,8 @@ class BladeAnonymousComponentTest extends TestCase
 {
     public function test_anonymous_components_with_custom_paths_can_be_rendered()
     {
-        Blade::anonymousComponentPath(__DIR__.'/anonymous-components-1', 'layouts');
-        Blade::anonymousComponentPath(__DIR__.'/anonymous-components-2');
+        Blade::anonymousComponentPath(__DIR__.'/Fixtures/anonymous-components-1', 'layouts');
+        Blade::anonymousComponentPath(__DIR__.'/Fixtures/anonymous-components-2');
 
         $view = View::make('page')->render();
 
@@ -25,8 +25,8 @@ class BladeAnonymousComponentTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        Blade::anonymousComponentPath(__DIR__.'/anonymous-components-1', 'layouts');
-        Blade::anonymousComponentPath(__DIR__.'/anonymous-components-2');
+        Blade::anonymousComponentPath(__DIR__.'/Fixtures/anonymous-components-1', 'layouts');
+        Blade::anonymousComponentPath(__DIR__.'/Fixtures/anonymous-components-2');
 
         $view = View::make('layouts::app')->render();
     }
@@ -35,14 +35,14 @@ class BladeAnonymousComponentTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        Blade::anonymousComponentPath(__DIR__.'/anonymous-components-1', 'layouts');
-        Blade::anonymousComponentPath(__DIR__.'/anonymous-components-2');
+        Blade::anonymousComponentPath(__DIR__.'/Fixtures/anonymous-components-1', 'layouts');
+        Blade::anonymousComponentPath(__DIR__.'/Fixtures/anonymous-components-2');
 
         $view = View::make('panel')->render();
     }
 
     protected function defineEnvironment($app)
     {
-        $app['config']->set('view.paths', [__DIR__.'/anonymous-components-templates']);
+        $app['config']->set('view.paths', [__DIR__.'/Fixtures/anonymous-components-templates']);
     }
 }
