+++ tests/Integration/View/BladeTest.php
@@ -245,10 +245,10 @@ class BladeTest extends TestCase
 
     public function test_view_cache_command_deduplicates_paths_before_compiling()
     {
-        View::addNamespace('templates', join_paths(__DIR__, 'templates'));
-        View::addNamespace('components', join_paths(__DIR__, 'templates', 'components'));
+        View::addNamespace('templates', join_paths(__DIR__, 'Fixtures', 'templates'));
+        View::addNamespace('components', join_paths(__DIR__, 'Fixtures', 'templates', 'components'));
 
-        Blade::partialMock()->expects('compile')->with(realpath(__DIR__.'/templates/components/panel.blade.php'));
+        Blade::partialMock()->expects('compile')->with(realpath(__DIR__.'/Fixtures/templates/components/panel.blade.php'));
 
         $this->artisan('view:cache');
     }
@@ -257,7 +257,7 @@ class BladeTest extends TestCase
     #[\Override]
     protected function defineEnvironment($app)
     {
-        $app['config']->set('view.paths', [__DIR__.'/templates']);
+        $app['config']->set('view.paths', [__DIR__.'/Fixtures/templates']);
     }
 }
