+++ tests/Integration/View/Fixtures/anonymous-components-1/app.blade.php
