+++ tests/Integration/View/Fixtures/anonymous-components-2/buttons/danger.blade.php
