+++ tests/Integration/View/Fixtures/anonymous-components-2/panel.blade.php
