+++ tests/Integration/View/Fixtures/anonymous-components-templates/page.blade.php
