+++ tests/Integration/View/Fixtures/templates/components/appendable-panel.blade.php
