+++ tests/Integration/View/Fixtures/templates/components/hello-span.blade.php
