+++ tests/Integration/View/Fixtures/templates/components/input-with-slot.blade.php
