+++ tests/Integration/View/Fixtures/templates/components/panel.blade.php
