+++ tests/Integration/View/Fixtures/templates/consume.blade.php
