+++ tests/Integration/View/Fixtures/templates/different-extension.sh
