+++ tests/Integration/View/Fixtures/templates/partials/scoped-partial.blade.php
