+++ tests/Integration/View/Fixtures/templates/uses-appendable-panel.blade.php
