+++ tests/Integration/View/Fixtures/templates/uses-include-regular.blade.php
