+++ tests/Integration/View/Fixtures/templates/uses-include-scoped.blade.php
