+++ tests/Integration/View/Fixtures/templates/uses-link.blade.php
