+++ tests/Integration/View/Fixtures/templates/uses-panel-dynamically.blade.php
