+++ tests/Integration/View/Fixtures/templates/uses-panel.blade.php
