+++ tests/Integration/View/Fixtures/templates/varied-dynamic-calls.blade.php
